Split database.py into it's own module
